Refactor elevations.js by extracting generateElevationSurfaces and hexToRgba to a utility folder
