--- california-dev/src/themes/light/utils/hexToRgba.js
@@ -1,7 +1,0 @@
-export function hexToRgba(hex, opacity) {
-  hex = hex.replace("#", "");
-  const r = parseInt(hex.substring(0, 2), 16);
-  const g = parseInt(hex.substring(2, 4), 16);
-  const b = parseInt(hex.substring(4, 6), 16);
-  return `rgb(${r}, ${g}, ${b}, ${opacity})`;
-}
